Updated part2 to work with FastRoute and Plates
This patch builds on #18, and updates it to work with FastRoute and Plates.

To make this work, I currently have it pinned to my feature/url-extension branch
of zend-expressive-platesrenderer, which adds the `url()` and `serverurl()`
functions to the Plates engine.

The only narrative changes were for the album list template; nothing else
changed.

For code changes, this includes the updated templates and configuration.